separated the users API (users.py) from the users forms (usersforms.py). separation of front-end and API will be important when we change front-ends from web to app.
